Added tests for clear and refresh.
